# Conflicts:
#	FiremanPro/.idea/misc.xml

+++ FiremanPro/database/src/main/java/com/project/test/database/controllers/AddressController.java
@@ -21,7 +21,7 @@ public class AddressController {
     public AddressController() {
     }
 
-    public Address addNewAddress(String streetName,String place, String streetNumber, Post post, long longitude, long latitude){
+    public Address addNewAddress(String streetName,String place, String streetNumber, Post post, double longitude, double latitude){
 
 Address address = new Address(streetName,place,streetNumber,post,longitude,latitude,CurrentDate,CurrentDate);
 address.save();
